Refactor gammapy download code base
